rename blogNav to topNav and use depject
