+++ app/page/blogIndex.js
@@ -7,7 +7,7 @@ exports.gives = nest('app.page.blogIndex')
 exports.needs = nest({
   'app.html.sideNav': 'first',
   'app.html.blogCard': 'first',
-  'app.html.blogNav': 'first',
+  'app.html.topNav': 'first',
   'app.html.scroller': 'first',
   // 'feed.pull.public': 'first',
   'feed.pull.type': 'first',
@@ -26,7 +26,7 @@ exports.create = (api) => {
 
     var blogs = api.app.html.scroller({
       classList: ['content'],
-      prepend: api.app.html.blogNav(location),
+      prepend: api.app.html.topNav(location),
       // stream: api.feed.pull.public,
       stream: api.feed.pull.type('blog'),
       filter: () => pull(
